Made the bitflag 64 bits.
git-svn-id: https://svn.forschooner.net/code/dnx/trunk/dnx@300 2be5450d-c085-40c6-83eb-9dc9c13f2dba

+++ common/dnxProtocol.h
@@ -148,7 +148,7 @@
 typedef struct DnxAffinityList
 {
    char * name;                     //!< Name of Nagios Host group/dnxClient
-   unsigned int flag;               //!< Flag for affinity check
+   unsigned long long flag;         //!< Flag for affinity check
    struct DnxAffinityList * next;   //!< Next structure in linked list
 } DnxAffinityList;
